+++ tos/consts.go
@@ -7,7 +7,7 @@ import (
 
 const (
 	// Version tos-go-sdk version
-	Version = "v2.6.1"
+	Version = "v2.6.2"
 )
 
 const TempFileSuffix = ".temp"